add validations and split checkout data
